Added security voters for each route

+++ src/Controller/DeleteFeature.php
@@ -4,6 +4,7 @@ namespace App\Controller;
 
 use App\Entity\Feature;
 use App\Repository\FeatureRepository;
+use App\Security\Voter\Verb;
 use Doctrine\ORM\OptimisticLockException;
 use Doctrine\ORM\ORMException;
 use Symfony\Component\HttpFoundation\Response;
@@ -24,6 +25,8 @@ class DeleteFeature extends Api
 
     public function __invoke(Feature $feature): Response
     {
+        $this->denyAccessUnlessGranted(Verb::DELETE, $feature);
+
         try {
             $this->featureRepository->delete($feature);
