+++ src/Controller/DeleteProject.php
@@ -4,6 +4,7 @@ namespace App\Controller;
 
 use App\Entity\Project;
 use App\Repository\ProjectRepository;
+use App\Security\Voter\Verb;
 use Doctrine\ORM\OptimisticLockException;
 use Doctrine\ORM\ORMException;
 use Symfony\Component\HttpFoundation\Response;
@@ -24,6 +25,8 @@ class DeleteProject extends Api
 
     public function __invoke(Project $project): Response
     {
+        $this->denyAccessUnlessGranted(Verb::DELETE, $project);
+
         try {
             $this->projectRepository->delete($project);
